third_party/WebKit/LayoutTests/fast/canvas/OffscreenCanvas-commit-frameless-doc-expected.txt

This is a testharness.js-based test.
PASS Verify that the getImageData() works on the OffscreenCanvas context of a frameless canvas
FAIL Verify that the placeholder canvas can be used as an image source assert_array_equals: property 2, expected 255 but got 0
Harness: the test ran to completion.
